+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## vNext
 ### Added
 - ThreadNumber is emitted in the result
-- Extension to log to log to the console 
+- Extension to log to the console 
+- Factory to create the IExecutionContext per Thread
   
 ## v2.2.1
 ### Fixed